requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
langchain-text-splitters
scikit-learn
streamlit
pypdf
python-dotenv